+++ pkg/amdgpu/gpuagent/gpuagent.go
@@ -68,6 +68,7 @@ type GPUAgentClient struct {
 	computeNodeHealthState bool
 	fsysDeviceHandler      *fsysdevice.FsysDevice
 	gCache                 *gpuCache
+	fl                     *fieldLogger
 }
 
 // Cache fields for GPUAgentClient
@@ -135,6 +136,7 @@ func NewAgent(mh *metricsutil.MetricsHandler, opts ...GPUAgentClientOptions) *GP
 
 	ga.healthState = make(map[string]*metricssvc.GPUState)
 	ga.mockEccField = make(map[string]map[string]uint32)
+	ga.fl = NewFieldLogger()
 	ga.rocpclient = rocprofiler.NewRocProfilerClient("rocpclient")
 	if ga.enableSriov {
 		logger.Log.Printf("profiler is disabled on sriov deployment")